opentelemetry-tracing/opentelemetry-python-tracing/flask-and-uswgi/requirements.txt
click==7.1.2
Flask==1.1.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opentelemetry-api==0.17b0
opentelemetry-exporter-jaeger==0.17b0
opentelemetry-instrumentation==0.17b0
opentelemetry-propagator-b3==0.17b0
opentelemetry-sdk==0.17b0
protobuf==3.15.0
six==1.15.0
splunk-opentelemetry==0.7.0
thrift==0.13.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
